+++ core/config/basicFungibleConfig.js
@@ -2,7 +2,7 @@ import harden from '@agoric/harden';
 
 import { noCustomization } from './noCustomization';
 import { makeCoreMintKeeper } from './coreMintKeeper';
-import { natExtentOps } from './strategies/natStrategy';
+import { natExtentOps } from './extentOps/natExtentOps';
 
 // Fungible tokens (our default for mints) do not customize
 // payments, purses, etc. They use the "basic" mintKeeper (the place
@@ -11,7 +11,7 @@ import { natExtentOps } from './strategies/natStrategy';
 // substraction and addition.
 
 // This configuration and others like it are passed into `makeMint` in
-// `issuers.js`.
+// `mint.js`.
 function makeBasicFungibleConfig() {
   return harden({
     ...noCustomization,
