+++ core/config/extentOps/natExtentOps.js
