+++ core/config/uniDescOpsConfig.js
@@ -2,7 +2,7 @@ import harden from '@agoric/harden';
 
 import { noCustomization } from './noCustomization';
 import { makeCoreMintKeeper } from './coreMintKeeper';
-import { makeUniExtentOps } from './strategies/uniStrategy';
+import { makeUniExtentOps } from './extentOps/uniExtentOps';
 
 // This UniDescOps config is used to create invites and similar assets.
 // It does not customize the purses, payments, mints, or assays, and
