+++ core/descOps.js
