+++ test/unitTests/more/imports/badImports.js
@@ -1,7 +1,7 @@
 // Copyright (C) 2019 Agoric, under Apache License 2.0
 
 import { importManager } from '../../../../more/imports/importManager';
-import { listIsEmpty, numIsEmpty } from './strategy';
+import { listIsEmpty, numIsEmpty } from './extentOps';
 
 function makeBadImportManager() {
   const mgr = importManager();
